fixes title sequence manager not updating
When going back to the main menu, the cursor position did not update, we
probably need to find a higher node that gets a new instance to fix
this, but for now this will update when you are on the screen.

Currently the value sticks because the memory does not change after you
confirm - i'll look into a way to fix this behaviour in the future.